cloud_storage: Log config in remote partition
At several points in remote partition such as creating a reader, waiting
for units, initializing a cursor, the config is logged. The client
address is part of this config so it should now be possible to take a
stuck connection and be able to trace its current state somewhat more
precisely than solely based on log offsets.