test: repro cstore crash
Normally, the segment meta store should be able to handle
segment replacements and out of order additions.

However, if we flush after every write we can effectively
skip all the logic in `column_store::insert_entries` and force
entries in the _head of the `column_store_frame`. When the frame fills
up (after 16 entries -- like below), it will "flush" itself to the
encoder.

The encoder for the `base_offset` column will receive out of order
data which it cannot handle and assert out.