transform-sdk: introduce Rust SDK
This is the initial take of a Rust SDK. The structure and APIs of the
Rust SDK closely mirrors the Golang SDK, except that due to Rust's
borrow checker, we have an explicit BorrowedRecord vs Record.

The SDK callback passes in BorrowedRecords but expects back Records so
that memory is managed correctly. We may way to do a similar approach
for the eventual C/C++ SDK.

An example of the identity transform is included in the repo, and it can
be compile to run in the broker via:
`cargo build --release --example=identity --target=wasm32-wasi`

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>