transform: guard against double stops
It's possible that a stopped processor (due to errors) is
stopped again because of shutdown or removal. This messes up the ref
counting in the cached engine. To guard against this we need to not call
stop twice on the engine.

Two options: Make the processor guard against double stops, don't call
stop twice in the processor. I think it's a the safer option to guard
against double stops in the processor, as the bookkeeping in the
processor is already fairly complex and this seems simpler.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>