security/jwt_test: Add AD-FS test data
Per usual:
- provider metadata
- JWKS example
- JWKS + JWT example
- Full authn data example

NOTE: The `sub` field here is constructed "by hand" to be the
application ID. AD-FS cum OIDC provider does not include a
subject in the client_credentials grant by default, but this
test demonstrates that it can be added by custom rule to satisfy
whatever requirements are in place on the broker.

Also includes a bunch of tooling-enforced formatting changes.