config: re-enable archival_metadata_stm consistency checks
These were disabled originally because it prevented older partitions
that had inconsistencies from bootstrapping properly, starting them in
incomplete states. It was made the default to ensure sweeping adherence
to "the old" data model, where inconsistencies were originally written
and applied to the metadata.

Upon disabling the checks, we ran into trouble applying certain
operations, sometimes resulting in a crash (see #14856).

So, this commit re-enables the consistency checks by default, with the
expectation that, when we do need to ignore inconsistent operations in
the log, we'll enable them only as needed.