transform: introduce txn_reader
Add a reader for transactions. This reader is able to filter out aborted
records and control batches so that the resulting stream of batches is
read committed isolation level (assuming the input reader only reads up
to LSO).

This reader is optimistic supporting a passthrough reading style until a
transaction is found, at which point we buffer all records into memory
filter out aborted/control records before continuing. This introduces
latency in the transform pipeline compared to the passthrough, but is
acceptable for transactions (the fetch path always reads all the records
from the reader before returning the fetch result).

I've spent a good deal of time reading the broker txn source and
franz-go's approach to handling transactions.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>