transform/tests/txn_reader: add tests for txn_reader
The tests for txn_reader are in two flavors.

The first one are standard unit tests that assert we treat
model::tx_range as an inclusive range. Special care is needed
here for comparing batches in GTest, as they are not copyable and
by default GTest only really supports copyable values in the matcher
framework. There is a workaround for this using a custom matcher that
unwraps std::reference_wrapper.

The second flavor of test is more a "fuzzer" style of test that takes in
a bunch of producer actions and then randomizes the order at which these
actions are applied. After applying those values it determines which
transactions should be committed vs aborted and compares that to the
output of the reader.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>