transform: use read committed isolation
Prior to this commit we used read uncomitted isolation and did not
suppress control batches to the WebAssembly Transform. Now we use the
new txn_reader to filter out aborted transactions and control records.

There are some requirements with lifetimes, namely that the reader takes
a partition_proxy, and we must ensure that the returned reader doesn't
outlive the source. This is already the case today, but some
documentation is added to make this more clear.

Also fix a bug where we forget to reset the gate after being stopped.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>