test: make ctest working directory configurable
This is used in the rp_unit to allow tests to be run in the current
working directory. This is pretty niche. The temporary in-memory
directory may be reasonable in >99% of use cases. However, some tests
require a "real" file system, for example expecting aio to have certain
behaviors related to alignment. The tmpfs behavior seems to just be cool
with whatever :(

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>