cloud_storage: Log client address in batch parser
The client address is added to the remote segment batch parser logs. The
address cannot be set as a member because a remote segment reader may be
cached and have its configuration reset. The client address is always
read from the configuration to make sure we get the latest address.