c/backend: limit initial learner start offset to manifest clean offset
When initializing learners newly added to the replica set we must be
certain that the locally retained data on the new replicas contains all
necessary segments for the Tiered Storage to work correctly.

Added additional limitation on a learner initial start offset by
clamping it with the manifest clean offset. This way a learner will
contain all the data that are required to continue archiving segments
without gaps.

Signed-off-by: Michal Maslanka <michal@redpanda.com>