r/state_machine: wait for the last snapshot to be applied
When state machine is applying the snapshot it may be the case that
another install snapshot request is received by underlying Raft protocol
instance. In this case the STM should immediately proceed to applying
the new snapshot instead of trying to read log from an offset which is
lower than the new log start offset.

Signed-off-by: Michal Maslanka <michal@redpanda.com>