transform: short circuit empty reads
This is actually required for correctness, as seen on the fetch path -
we request the prev_offset for LSO (see this commit as to why
https://github.com/redpanda-data/redpanda/commit/498031126c8f6b41e842b256e8df9fc1733bee6b).

However when topics are empty (LSO=0) this means that we request the max
offset to model::offset::min(), which translation gets upset about. We
fix in the same manner as the read path by short circuiting empty reads.
This is the same behavior as the fetch path (see this commit
https://github.com/redpanda-data/redpanda/commit/7aa02b5d7943fad6b35a54073de793c54bc30e21#diff-ccfd5e131023aa80d73e854c51f6dd29973c5fe5003c004aa4487f637319879dR185),
although I can see the argument that LSO=0 should translate to
max_offset=0 instead of relying on the short circuiting behavior.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>