c/archival_stm: translate sync error to not_leader error code
When `persisted_stm::sync()` method fails it is indicating that the
current node is not longer a leader. The `sync()` executed before
`replicate` call in archival stm `command_batch_builder` prevents
replicate from being called. The end result for such an error is
deterministic and we can translate the sync error to `not_leader` error
code.

Fixes: #14898

Signed-off-by: Michal Maslanka <michal@redpanda.com>