ducktape: use latest seqno instead of counts
It's possible that transforms don't start processing until after the
first couple of records are produced on the input topic. Since we start
processing at the end of the log, we can skip records and taht causes
the test to fail because it never reads the initial records. Instead of
checking progress based off of record counts, use the seqno that we
use to verify there are no skipped records.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>