storage: consider segments with only non-data batches as compacted
When the final segments in a sliding range contains only non-data
batches, we could previously end up recompacting the same range over and
over again. Consider the given sliding range, where segment [20, 25) is
all non-data batches.

[0, 10)[10, 20)[20, 25)

The compacted indices only contains data keys, so the key offset map
will only reflect data up to offset 20. Because of this, we would skip
compacting [20, 25) and leave the segment as is.

Subsequent attempts to compact would see that there is an uncompacted
segment, and attempt to compact the sliding range again, thinking that
the uncompacted data is new data to be merged in.