ducktape/cloud_storage: Increase producer msg count
During the model test the expectation is that at least seven or eight
segments will be downloaded for TS_Read. The test produces 10MiB of
data, 1MiB segment size and sets 3MiB as local retention size. The
actual segments uploaded are roughly 2MiB each, resulting in fewer
downloads.

This change doubles producer message count so that we need to produce
and later download more segments to consume.