rptest: keep track of expected anomalies
Previously, the scrubbing test did not keep track of the anomalies that
are expected throughout the test. Instead anomalies were always compared
against the previous ones. This turned out to be problematic because in
most cases a missing segment or a missing spillover manifest will also
be recorded as an offset gap. One could run into failures where the
offset gap was detected later on, which caused the comparisons described
above to fail.

The fix is to track the expected anomalies throughout the test and check
that these expected anomalies are a subset of the newly reported
anomalies.