cloud_metadata: add producer_id_recovery_manager
Adds an entity that orchestrates restoration of the producer_id. It
first sends RPCs to each node to determine the highest producer ID that
exists in archival segment data, and uses an RPC directed at the
id_allocator leader replica to reset its next producer_id.

This will be plugged into the cluster_recovery_backend.