metrics: Add test to check for expected metrics
When 23.2.1 got released originally we ran into a metric explosion
because of some newly added metrics that had too high cardinality. We
want to avoid such situations in the future.

It's easy to add (too many) metrics. Hence this patch adds a DT test
which makes adding metrics explicit. The test spawns a RP cluster and
scrapes the metrics followed by a comparison against a checked in golden
metrics file.

The idea is not to discourage adding new metrics but to make the process
explicit and force people to think about the cardinality of the metrics
that are being added.

We further provide guidelines of what best practices are in terms of
adding metrics: https://vectorizedio.atlassian.net/l/cp/F8sAaV6D

We check three different modes:
 - internal metrics - aggregation off
 - internal metrics - aggregation on
 - public metrics

To make it easy to update the golden metric files we add another "test"
which can be run like any other DT test to regenerate the golden metrics
file. By default this doesn't do anything (so as not to introduce any
implicit ordering requirements with the actual test).

The test is disabled on CDT as the amount of metrics created in
dedicated DT mode is slightly different. This isn't much of an issue as
the main purpose for this test is to run on branch builds and detect
changes at change time.

Note in a similar vein this test will never be able to catch all metrics
as some are created dynamically (which is not a recommended prometheus
practice). It's not the end of the world, we do create topics and
partitions which is where most of the metrics explosions come from.

Note we discussed whether we should make the test "smart" in the sense
that it would allow differences that are always fine (e.g.: adding a
single per shard counter). However we decided against that as it would
introduce unnecessary diffs when later being forced to update the golden
files.  Further given updating the golden metrics files is fairly easy
it's good to always be explicit about this.