r/recovery_stm: stop recovery when follower was already updated
Recovery and replicate stms are not synchronized. It may be the case
when both of stms are active at the same time that the same batch is
delivered to the follower twice. In general this batch duplication is
harmless as Raft is not vulnerable for messages redelivery but it may
cause unnecessary truncation and latency increase.

Added a check validating expected log end offset right before sending
recovery append entries request. This will prevent sending the same set
of batches twice to the follower.

Fixes: #14413

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 7bb54d3c5e747f0809e7d9fc74a36e2a8a34e25a)