transform: enable processor to control lifecycle
Currently, when the manager is responsible for marking when a processor
is active. Move this responsiblity to the processor so that we can delay
when we mark a processor as "started" until the initial offset has be
read (see the next commit).

A future change in this direction is moving the start/stop processor
work to be complete async if they actually end up being expensive enough
to block other work.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>