transform: mark start after offset load
We want to wait to mark processors as running until we've loaded the
initial offset. The majority of the latency here is the delay to wait
for previous processors to start. Since doing that work in `start` would
block more work from happening in the manager, there is a desire to make
this async.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>