rptest: use exclusive ranegs in scrubbing test
Previously, the code that creates anomalies could pick the same offset
range to insert anomalies in. This is problematic for validation of
metrics since it may mean, for example, that a gap won't be recorded
since it's in the offset range of a missing spillover manifest already.

Note that the `BucketView.is_segment_part_of_a_manifest` has been
renamed to `BucketView.find_segment_in_manifests` and instead of
returning a boolean result, it will return the segment meta if the
segment was found and None otherwise. Call sites have been updated
accordingly.