wasm/alloc: zero fill heap memory
wasmtime assumes that memory is zero filled, so ensure that heap memory
is zero filled. We also force callers to tell the allocator how much of
that heap was actually used, so that deallocation can memset only the
changed memory - this is an optimization for large memories if only part
of the memory is used.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>