features: do not retire serde_raft0 feature
The `serde_raft0` feature can not be completely retried from the feature
table as the feature presence is checked in the previous versions of
Redpanda. We need to first remove the checks validating if feature is
active and only then retire the feature completely in the next version.

Retiring a feature caused the upgrade test to fail as the feature is
expected to be active when some other subsequent features are active.

Fixes: #15046

Signed-off-by: Michal Maslanka <michal@redpanda.com>