build: trigger github action build on "ready for review"
By default pull_request event is triggered only for opened, synchronize,
and reopened activity types. If a PR is opened in a draft state, then
transition to ready for review, the build won't be triggered again. To
fix this, we define explicitly the list of activity types for which the
build needs to be triggered and include "ready_for_review" to the list.

References:
- https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#pull_request