rptests: Allow latency threshold adjustment
This change allows the base latency thresholds for the various
percentiles in the OMB verification test to be adjusted by a multiplier.

We will use this ability to relax the latency thresholds for tests we
know are slower because, for example, there max out one of the supported
resources dimensions.