broker: Use byte-granular node memory size in check_cluster_limits()
To see how close to the limits we can push the number of partitions
based on node memory, we need accurate memory stats. Today we only have
GiB-granular sizes which means if a node has some fractional GiB value
(a common scenario in the cloud) we're leaving partitions on the table.

In the worst case, if a node has 15.9GiB of memory we're only able to
allocate 15GiB without triggering the checks in check_cluster_limits(),
wasting 900MiB+ or 249 partitions if topic_memory_per_partition=4MiB.

Add a new 'available_memory_bytes' field to broker_properties to make
these checks more accurate.