archival: Use new types for candidate creation
The ntp archiver relies on tools in archival policy to create upload
candidates. These are modified to return errors during candidate
creation explicitly, instead of inferring that creation failed from
upload candidate members.

It is possible to further propagate the type up the call chain so that
we do not use empty upload candidates when waiting for uploads to
finish, but this can be addressed in a different changeset.

(cherry picked from commit 2512d4ee2d4989a2fe648782c692cdfacdf89367)