tx/migrator: added tx manager migrator service
Added a service handling tx manager topic migration (rehashing of tx
manager stm updates to new partition count). A service is a simple state
machine that works in the following steps:

1) Create a temporary topic for new partitioning scheme
2) For each partition of current tx manager topic read all the data,
   assign new partition and replicate according to the new scheme
3) Delete current tx manager topic
4) Create tx manager topic with new number of partitions
5) Replicate data from temporary topic to the new tx manager topic.
6) Delete temporary topic

If the process is interrupted with a failure it may be retried as the
logic in migrator service determines starting condition and adjust
initial step accordingly.

The error handling is very simple and requires retrying operation if it
failed.

If any of the operation failed it will always be started from scratch to
prevent leaving some intermediate state behind.

Signed-off-by: Michal Maslanka <michal@redpanda.com>