archival: fix collector when range start is inside segment
Consider the following local and cloud log layout:
      +-----------------------------+------------------+---------------+
Local |0                         999|1000          1099|1100       1199|
      +-----------------------------+-----------+------+---------------+
      +--------------------+--------+------------------+---------------+
Cloud |0                989|990  999|1000          1099|1100       1199|
      +--------------------+--------+------------------+---------------+

Asking the `segment_collector` to collect the [990-1199] range is a
perfectly valid thing to do. Previously, this collection would fail.
The segment collector does it's own size accounting of the range from
local disk and tries to respect the max size provided in the
constructor. If the requested range starts inside a segment, then it
would previously account for the size of the entire segment. In the
example above, we'd account for the entire size of the first segment
even though we only needed the last batch. This would lead to running
out of size before reaching the end of the range.

This commit proposes a hacky solution where we treat the first segment
as free if the offset range to collect starts within it. I also
considered determining the required size precisely, but that requires
turning a lot of sync code into async.