vassert: split header and implementation in three libs
this allow at link time to choose what happen when a vassert fail

the redpanda implementation is to terminate the program, and it's
implemented in v::vassert_abort_implementation - vassert.c

the alternative implementation is to launch an exception with backtrace,
and it's implemented in v::vassert_exception_implementation -
vassert_as_exception.cc. this alternative is useful in fuzzing tests,
where the terminating assert makes more difficult to pinpoint the
failure.

at linking level, v::vassert_header is linked to every v_cc_library.
then the implementation is selected when declaring the executable.