transform/processor: fix off by one lag error
This fixes an off by one error with lag. _source->latest_offset() is LSO
which is last_produced_offset + 1, we want to compare this with the last
produced offset.

NOTE: that latest_offset can return 0 if there is an empty log, but we
don't have to worry about that here because if the transform read
something the log isn't empty :)

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>