archival/test: Ensure archiver is stopped
If the test affected by this change fails, archiver is destroyed without
stopping. This also destroys the scrubber, which may have requests
active keeping its gate open.

This triggers an assertion from seastar related to gate destroyed while
requests active. Usually this is prominently visible but in this test it
results in the ctest invocation hanging in CI.

This causes the build to hang and timeout. The root cause is the
test_manifest_spillover not adding segments to the archival meta STM,
which can be investigated separately. This change should enable the test
to fail in a visible manner with logs seen in CI instead of hanging.

(cherry picked from commit ef8a4203643b9402fbfc5d0003c5dd7b19fb16d3)