build: trim cache to avoid accumulation
The github build action uses ccache to speed up builds. Before a run a previous snapshot of the cache is restored, and after the run a new replacement snapshot is taken.

This works great, but has the side effect that the cache grows, each time accumulating new bits along side old bits. A fresh build for example might be around 190M of cache use, but I see that we are now saving cache snapshots >1G.

So instead we'll trim down before taking the snapshot. Generally we may want to periodically check-in and see if the cache size limit is large enough for a fresh build.