rptest: accommodate cluster metadata uploads in tests
The read replica clusters in tests are typically started without
creating a bucket for the cluster, which yields some error logs
complaining about missing buckets when configuring to upload cluster
metadata. So, this disables uploading cluster metadata in most of these
tests. I did the same for other tests that play around with bucket
configs.

The upgrade test is a special beast: there isn't an easy wa to have the
nnew cluster start up with a non-default config. So, I opted to create
a bucket for the read replica in this case.