c/tm_stm: clear cache on tm_stm startup
As tm_stm owns the tm_stm_cache it should clear it every time we start a
new instance. This way we guarantee that the state will be managed
solely by the tm_stm instance that is currently running.

For now we do not want to clear the state on shutdown as we may start
seeing UNKNOWN_SERVER_ERRORS caused by partition movements. (right now a
node can always ask others about the previous tx state).

Signed-off-by: Michal Maslanka <michal@redpanda.com>