tests: made progress timeout dependent on failure injection
Made progress timeout longer in the tests that enables failure
injection. With random failure injection it may be the case that the
node which is currently controller leader will be restarted/suspended
multiple times before being able to finish balancing operation. In this
pessimistic case it may happen that the tick will not be scheduled
frequent enough for the decommission to make progress.

Fixes: #8657

Signed-off-by: Michal Maslanka <michal@redpanda.com>