storage: deduct compaction memory from memory groups
The default behavior of windowed compaction is to allocate 128MiB per
shard and maintain that for the duration of the node, reusing it across
compactions. This can be problematic for memory-constrained deployments,
e.g. in CDT nodes we see nodes with two shards and ~350MB, and
compaction takes a majority of this memory.

This commit carves out a chunk of memory, reserving it for compaction,
disallowing use by other subsystems.