resource_mgt: restrict compaction memory based on total memory
The default behavior of windowed compaction is to allocate 128MiB per
shard and maintain that for the lifetime of the node, reusing it across
compactions. This can be problematic for memory-constrained deployments.
E.g. in CDT we see nodes with two shards and ~350MB, and compaction
takes a majority of this memory. While artificial, similar environments
are sure to exist when trialing Redpanda, and we should make sure to
avoid crashing the system for a feature that may not always be used.

This commit imposes a configurable percent limit on the amount of memory
to be used by the compaction hash key-offset map. When configured, the
min between the calculated bytes value and
`storage_compaction_key_map_memory` is used to reserve compaction
memory.