cloud_metadata: recover remote topic properties from controller snapshot
We previously relied on downloading the topic manifest to be able to
restore a topic. This meant that for cluster recovery, there would be a
window between uploading the controller snapshot and uploading the first
topic manifest where a topic would cause cluster recovery to fail.

This commit remediates this by building the remote_topic_properties with
content from the restored constroller snapshot. A corresponding change
is made to the topics_frontend such that when remote_topic_properties is
provided, we skip the step to attempt downloading the topic_manifest,
instead passing through the provided remote_topic_properties.