kafka: skip restoring groups that already exist
The current implementation of consumer offsets recovery be always
overwrite existing groups. This makes the operation idempotent, but also
a bit risky cases where an application begins committing new offsets
before the restore process is complete, as newly committed offsets could
be lost.

This commit updates the behavior of group_manager::recover_offsets() to
skip over groups that already exist and are in a non-empty state.