r/consensus: made leader_id and term changes atomic
Fixed an error in Redpanda Raft protocol implementation leading to
reporting inconsistent leadership information. Previously it may
happened that the protocol `_term` was updated but `_leader_id` was
not. This might lead to situation in which a caller querying `term()`
and `leader_id()` method would receive new `term` value but the old
`leader_id`. The problem is limited to metadata and does not influence
protocol correctness as all the protocol checks are correct. The issue
might have been underlying some tests failures as the incorrect
leadership information might have propagated through health report
leading to leadership metadata staleness.

Signed-off-by: Michal Maslanka <michal@redpanda.com>