cloud_metadata: be more strict about recovery stage order
We could previously see recovery stages go "backwards" in a benign way,
if we previously had replicated the stage change, changed leadership,
and then reconciled the controller snapshot only to discover we need to
reapply the cluster config again (we don't diff the config against the
current config, we just apply it whole).

This updates the reconciliation code to consider the current stage when
determining what steps need to be performed, allowing us to more easily
reason about the contents in the controller log.

A couple of test cases are updated to perform such validation.