storage: don't recompact segments with no compactible data
Segments may contain records that won't be compacted:
- when all batches are non-data batches, or
- when the only data record is the last record

In either of the above cases, a windowed compaction will repeatedly
result in the same resulting segment, since we only remove non-data
batches and we do not remove the last record.

This commit introduces a new field to the index_state to keep track of
the first data batch in the segment. This is used to determine whether
either of the conditions are true before compacting them. If so, the
segment is skipped for compaction.

Fixes #15347