r/consensus: update _last_leader_visible_offset on leader as well
Previously _last_leader_visible_offset was updated only from
info in append_entries requests. This meant that on a leader that was
stepping down there was a small window between stepping down and receiving
the first append_entries from the new leader, where
_last_leader_visible_offset value was bogus. This led offset_out_of_range
errors when serving valid fetch requests as a follower. Fix this by
updating _last_leader_visible_offset value while we are the leader as
well.