rpk produce: Adjust timeout and maxBytes options
1. ProduceRequestTimeout: we are now using the
duration set by `--delivery-timeout` to also
adjust the ProduceRequestTimeout. We are using the
same flag to avoid adding another knob that may
lead to confusion.

2. BrokerMaxWriteBytes: we were only adjusting
the ProducerBatchMaxBytes with the
`--max-message-bytes`, that means we were only
maxing out batch bytes but not the broker max
bytes and we would likely hit the broker byte
batch limit if not increased.is limit if we were
producing a record big enough.