config: Add hydration timeout
A 10 minute hydration timeout is added. If the hydration request is not
completed within this time period, the consumer is notified with a
timeout error. The download which happens in the remote segment via a
condition variable may still progress, but the connection to the
consumer can be closed.

(cherry picked from commit 80958b1c641bb3d5b10c2fe1ef48c627a4ed74c3)