cloud_storage: Make hydration-wait abortable
A consumer may disconnect while waiting for hydration. The fiber waiting
for hydration in redpanda should be able to react to this and stop
waiting, thus enabling the kafka connection to close. Additionally, the
hydration may get blocked somewhere for long period of time, in which
case it is useful to be able to abort the wait for hydration.

This change allows the fiber waiting for the hydration to complete to
abort if the consumer disconnects or there is a timeout. In either case
the download continues uninterrupted, only the consumer is allowed to
close the fetch path early.

(cherry picked from commit dba77b0f4b00a00cc2eb8ea6ff3ee34488573cd0)