storage: fix tracking first data record on truncation
When a segment is suffix truncated, the first_data_offset record of its
index needs to be invalidated. This value is used to determine whether
the first data offset is the last record, in which case we know a
compaction will be a no-op. If the first data offset is made incorrect,
this check may be incorrect and we'd end up doing a needless compaction.