storage: compaction short-circuit to consider compactible data
The current implementation of a short circuit for compaction looks at
the offset of the first data record, and uses it to infer whether the
segment is all non-data records or has one data record at the end of the
segment. In either case, a compaction will be pointless since the
segment size will not be reduced.

This is slightly too strict a condition: compaction won't gate
compaction of a batch on user-data or not, but rather based on a set of
record batch types. This updates the field and associated methods to
account for this.