storage: start compaction sliding range if segment has multiple data records
When the front of the compaction range is all non-compactible batches,
we don't need to bother considering them for compaction. This is the
case when there's a small set of keys that get frequently updated.

This commit updates the logic in finding the sliding range to skip over
these initial segments.