tx/compaction: tighten is_transactional check
For transactional batches we invalidate the compaction index so it is
later built during compaction time. When checking transactionality we
are missing transactional control batches which _do not_ have  the
transactional bit set but rather have a pid and control bit set. Fix
the condition so compaction index is invalidated correctly.

This is a rare edgecase that was caught by random transaction generator.