c/archival_stm: do not reset _last_replicate on timeout
For same term syncing we store the replicate commands future so that it
can be awaited in the next sync call. However, the
`std::exchange(_last_replicate, std::nullopt)` before the wait is
problematic as it "forgets" the last replicate. If sync times out and we
retry it then it behaves as if the last replicate command succeeded.
This is not necessarily true as the newly added test assertion shows.

Use a shared future for _last_replicate so that it can be awaited
multiple times and reset it only after it is resolved. This guarantees
that sync will only return after last replicate command actually
resolved.

`ignore_ready_future` call is removed as the shared future does not
issue a warning if it wasn't consumed (basically a revert of
e221a0aab8).

(cherry picked from commit 299e12dfb241df84068395c1568f357a1e16ddda)