r/persisted_stm: remove persistent state from previous incarnations
This is particularly relevant for kvstore-backed stms. They use ntp
without revision as the kvstore key, therefore if there is leftover
state from the previous partition instance (e.g. topic was re-created,
and we relied on orphan log dir deletion to remove the previous
instance), it could erroneously get used for the new partition instance.
Delete it before the first start.