storage: still self compact one-record segments
A previous commit changed the range over which to do window compaction
to ignore segments at the beginning of the log with only one records.
Some tests expect that segments will at least be self compacted, since
this is what rebuilds indexes, handles transactions, etc.

So, this commit tweaks the logic to still keep these segments in the
sliding range initially when performing self compaction, but then before
running windowed compaction, removing those in the beginning that may
not have comapctible records.