Add backtrace: in Top-N alloc site output
When we OOM we output the top-N alloc sites and a one-line backtrace.

However, sesatar-addr2line fails to recognize it because the one-line
regex requires backtrace: or at: in order to understand that it's a
one-line backtrace.

So we add 'at:' in the string, which is a recognized token to
introduce a backtrace.

Fixes #15615.

(cherry picked from commit d7547d49cfa87e38c2139afa38b0701b873f47f2)