test: fix build type switch in upgrade tests
In a docker runner head may point some where based on a previous run
(e.g. the release build type):

  [nwatkins@fedora clang]$ docker exec docker-rp-1 ls -l /opt/redpanda_installs/
  total 0
  lrwxrwxrwx. 1 root root  68 Jul 14 16:04 head -> /home/nwatkins/src/redpanda/vbuild/release/clang/dist/local/redpanda
  drwxr-x---. 8 root root 101 Sep 29 18:41 v22.1.11

But that might not exist on the host:

  [nwatkins@fedora clang]$ docker exec docker-rp-1 ls -l /opt/redpanda_installs/head/
  ls: cannot access '/opt/redpanda_installs/head/': No such file or directory

Furthermore, in the test, it doesn't get updated to point at the correct
build type. Above was examined after running with build_type=debug

  --globals='{"rp_install_path_root":"/home/nwatkins/src/redpanda/vbuild/debug/clang/dist/local/redpanda"

And then in tests we get failures like this

  nohup: failed to run command '/opt/redpanda/bin/redpanda': No such file or directory

The issue is that the symlink for head version path wasn't being updated
if it existed.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>