cloud_storage_clients: tolerate spurious cvar wakeups
The test failed randomly because the pool semantics were properly
enforced. A condition variable can wakeup without being signaled which
led to connection borrowing when it wasn't supposed to happen (per
test).

The pool semantics is roughly this: lease http clients. If locally
managed pool is empty, borrow from peer shards if possible up to a
limit. If we can't or are not allowed to borrow, wait until clients are
returned to the local pool.

Note: the purpose of this semantics isn't exactly clear to me, I'm just
fixing the code to work per test.