cloud_metadata: upload consumer groups in batches
We currently serialize the entirety of the consmer offsets partition
into a single iobuf. This imposes a fundamental limit on the scale of
the number of groups that can be snapshotted.

The cluster manifest previously accounted for this by allowing multiple
snapshots per offsets partition in the manifest, even though existing
callers would only ever pass a single snapshot for the entire partition.

This commit begins batching the group snapshots with a configured value
for the number of groups.

(cherry picked from commit 3113ca885ef229d58aae4d7dc0393b6b02ad9129)