config: property for changing consumer offsets upload batch size
This introduces a safeguard config, protecting against uploads from
consumer offsets partitions that may countain a huge number of groups.
As is, the offsets uploader will upload all groups as a single iobuf,
which imposes an evental limit on the scalability of offsets uplaods.

This commit introduces a property to cap the number of groups per
snapshot that will be used in a subsequent commit.

(cherry picked from commit 847aaae16bf277406779f30e1011d2cd6641254f)