rptest: clean up cluster metadata in CDT
In CDT we examine each bucket hash and use that to determine the list of
objects to delete. This wasn't taking into account the new cluster
metadata prefix. The result is that we would not completely empty the
bucket before attempting to delete it, resulting in a failed bucket
deletion.