rptest: log error on failure to delete bucket
In general this seems like an unexpected case. Perhaps it's sometimes an
S3 issue, but it's certainly possible this is caused by a ducktape bug
(see https://github.com/redpanda-data/redpanda/pull/15754).

In either case, it seems worthwhile to flag, if anything, to inform us
when to expect artifacts to show up in our S3 account.