storage: account for windowed compaction in backlog computation
The current backlog computation doesn't account for IO done by windowed
compaction: in addition to reading all segments for self compaction, we
read them again post-self compaction once they are window compacted.

This adjusts the calculation to account for this by determining the
sliding range and scaling the to-be-window-compacted segments by the
compaction ratio.

Fixes #15756