k/fetch: track bytes read from the follower with metric
Since it is hard to validate if Follower Fetching feature is working
correctly we add a separate `bytes_fetched_from_follower_total` metric
that counts bytes that were read from the follower. This will make it
very easy to validate if consumers are reading from the closest replica.
Note: follower fetched bytes are also accounted to overall bytes read.

Signed-off-by: Michal Maslanka <michal@redpanda.com>