Fix iobuf zero-copy for full fragments
When appending one large iobuf to another (or a large temporary_buffer
to an iobuf) one would expect that this is mostly a zero-copy operation,
 as the underlying temporary buffers can be shared and iobuf is designed
to accommodate this case.

However, a check in the existing logic which tries to avoid creating
many small iobufs also kicks in when we append a series of full-sized
fragments to the iobuf, resulting in a full copy of each of them.

This change is a relatively conservative change to this logic where
we will now do zero-copy when a full-sized (or larger) temporary
buffer is appended. We don't otherwise change the heuristic though one
could argue that a future improvement where we also zero-copy some
"large but not totally full" buffers would be beneficial.

I also remove a redundant check in this heuristic:

b.size() < available_bytes()

as it is subsumed by the next check.

Partial fix for issue #15769.

(cherry picked from commit 3953fb5bf9c8374094bc358cb671d3c2d747d892)