c/rm_stm: cleanup tx_locks
Currently tx_locks are never cleaned up and the _tx_locks map grows
unboundedly. Fix this by removing the lock when we cleanup the producer
state. I'm not sure if we can remove the lock (optimistically?) when
transactions are aborted or committed, as we erase some in memory state
at that point. I'm being more catious and not doing that in this PR, but
it seems like if nobody is holding the lock at the time when we
abort/commit a transaction, we can clean the lock up, as it's only there
to prevent concurrent tx related requests for a single pid.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>